+++ Market/Entities/Storage.cs
@@ -11,13 +11,13 @@ namespace Market.Entities
         public int ID { get; set; }
         public int DispatchNoteID { get; set; }
         public string Barcode { get; set; }
-        public float PriceForUnit { get; set; }
-        public int Amount { get; set; }
+        public double PriceForUnit { get; set; }
+        public double Amount { get; set; }
         public List<Product> Products { get; set; }
         public Storage()
         {
         }
-        public Storage(int DispatchNoteID, string Barcode, float PriceForUnit, int Amount)
+        public Storage(int DispatchNoteID, string Barcode, double PriceForUnit, double Amount)
         {
             this.DispatchNoteID = DispatchNoteID;
             this.Barcode = Barcode;
